Added optional args to  unset-commit-template.sh
- unset-commit-template.sh -D will skip deletion of template file
- Added documentation to the headers of both commit template scripts

+++ unset-commit-template.sh
@@ -3,9 +3,20 @@ set -o errexit
 # ==============================================================================
 # unset-commit-template.sh
 # Author: Connor de la Cruz (connor.c.delacruz@gmail.com)
+# ------------------------------------------------------------------------------
+# Unset local git config for commit.template if configured. Template file will
+# be deleted unless -D argument was specified.
 #
-# TODO DOCUMENT:
-#   - usage
+# ------------------------------------------------------------------------------
+# Usage
+# ------------------------------------------------------------------------------
+# Running unset-commit-template -h will display details on usage and arguments:
+#
+# Usage: unset-commit-template.sh [-D] [-h]
+# Options:
+#   -D  Don't delete commit template file.
+#   -h  Show this help message and exit.
+#
 # ==============================================================================
 
 # Functions --------------------------------------------------------------------
@@ -28,15 +39,41 @@ git_commit_template() {
     git config --local --get commit.template
 }
 
+# Display help message for this script
+show_help() {
+    echo "Usage: unset-commit-template.sh [-D] [-h]"
+    echo "Options:"
+    echo "  -D  Don't delete commit template file."
+    echo "  -h  Show this help message and exit."
+}
+
 # Main -------------------------------------------------------------------------
 
 # Checks the local commit.template config for the current repo. If configured,
 # will unset and delete the template file it was set to.
+# Arguments:
+#   Takes all optional arguments for script. For details on these arguments,
+#   see show_help()
 main() {
     # Check that this is a git repo
     verify_git_repo
 
-    # TODO: -D to not delete file?
+    # Parse arguments:
+    # -D (don't delete template file)
+    local arg_no_delete
+    while getopts 'Dh' opt; do
+        case ${opt} in
+            D)
+                arg_no_delete=1
+                ;;
+            h|?)
+                show_help
+                [[ "$opt" = "?" ]] && local exit_code=1 || local exit_code=0
+                exit $exit_code
+                ;;
+        esac
+    done
+    shift $((OPTIND -1))
 
     # Get template (if configured)
     local commit_template_file="$(git_commit_template)"
@@ -45,17 +82,19 @@ main() {
     echo "Unsetting commit.template..."
     git config --local --unset commit.template
 
-    echo "Removing template file..."
-    local current_dir="$(pwd)"
-    local repo_root_dir="$(git_repo_root)"
-    cd "$repo_root_dir"
-    rm -f "$commit_template_file"
-    echo "Template removed."
-
-    # Return to previous directory before exiting
-    cd "$current_dir"
+    if [[ -n "$arg_no_delete" ]]; then
+        echo "-D was specified, leaving template file $commit_template_file."
+    else
+        echo "Removing template file..."
+        local current_dir="$(pwd)"
+        local repo_root_dir="$(git_repo_root)"
+        cd "$repo_root_dir"
+        rm -f "$commit_template_file"
+        echo "Template removed."
+        # Return to previous directory before exiting
+        cd "$current_dir"
+    fi
 }
 
 # Run main, pass any command line options to it for parsing
 main "$@"
-
